% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjusted_efficiency.R
\name{get_adjusted_efficiency}
\alias{get_adjusted_efficiency}
\title{get_adjusted_efficiency}
\usage{
get_adjusted_efficiency(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Filters results to the specified season.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns adjusted offensive and defensive efficiency ratings.
}
