% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_picks.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{get_draft_picks}
\usage{
get_draft_picks(
  year = NULL,
  draft_team = NULL,
  source_team = NULL,
  position = NULL
)
}
\arguments{
\item{year}{Filters results to the specified draft year.}

\item{draft_team}{Filters results to the specified NBA team.}

\item{source_team}{Filters results to the specified college team.}

\item{position}{Filters results to the specified player position abbreviation.}
}
\description{
Returns historical NBA draft picks.
}
