% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elo.R
\name{get_elo}
\alias{get_elo}
\title{get_elo}
\usage{
get_elo(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Filters results to the specified season.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns historical Elo ratings.
}
