% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_games}
\alias{get_games}
\title{get_games}
\usage{
get_games(
  start_date_range = NULL,
  end_date_range = NULL,
  team = NULL,
  conference = NULL,
  season = NULL,
  season_type = NULL,
  status = NULL,
  tournament = NULL
)
}
\arguments{
\item{start_date_range}{Includes games starting at or after this ISO 8601 timestamp.}

\item{end_date_range}{Includes games starting at or before this ISO 8601 timestamp.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}

\item{season}{Filters results to the specified season.}

\item{season_type}{Filters results to the specified season type.}

\item{status}{Filters results to the specified game status.}

\item{tournament}{Filters results to the specified tournament, such as NCAA or NIT.}
}
\description{
Returns up to 3,000 games that match the filters, ordered by start date.
}
