% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lines.R
\name{get_lines}
\alias{get_lines}
\title{get_lines}
\usage{
get_lines(
  season = NULL,
  team = NULL,
  conference = NULL,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{Filters results to the specified season.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}

\item{start_date_range}{Includes games starting at or after this ISO 8601 timestamp.}

\item{end_date_range}{Includes games starting at or before this ISO 8601 timestamp.}
}
\description{
Returns betting lines for up to 3,000 games that match the filters, ordered by start date.
}
