% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineups_by_team_season.R
\name{get_lineups_by_team_season}
\alias{get_lineups_by_team_season}
\title{get_lineups_by_team_season}
\usage{
get_lineups_by_team_season(
  season,
  team,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{The season to return.}

\item{team}{The team name to return.}

\item{start_date_range}{Includes games starting at or after this ISO 8601 timestamp.}

\item{end_date_range}{Includes games starting at or before this ISO 8601 timestamp.}
}
\description{
Returns lineup statistics for a team and season.
}
