% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{get_overview}
\usage{
get_overview(team_id, season)
}
\arguments{
\item{team_id}{Team ID filter}

\item{season}{Season filter}
}
\description{
Returns a stored full-season team overview
}
