% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_season_shooting_stats.R
\name{get_player_season_shooting_stats}
\alias{get_player_season_shooting_stats}
\title{get_player_season_shooting_stats}
\usage{
get_player_season_shooting_stats(
  season,
  season_type = NULL,
  team = NULL,
  conference = NULL,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{The season to return.}

\item{season_type}{Filters results to the specified season type.}

\item{team}{Filters results to the specified team name. Required when conference is not provided.}

\item{conference}{Filters results to the specified conference abbreviation. Required when team is not provided.}

\item{start_date_range}{Includes games starting at or after this ISO 8601 timestamp.}

\item{end_date_range}{Includes games starting at or before this ISO 8601 timestamp.}
}
\description{
Returns player shooting statistics for a season. Provide a team or conference.
}
