% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays.R
\name{get_plays}
\alias{get_plays}
\title{get_plays}
\usage{
get_plays(game_id, shooting_plays_only = NULL)
}
\arguments{
\item{game_id}{The game ID.}

\item{shooting_plays_only}{When true, returns only shooting plays.}
}
\description{
Returns all recorded plays for a game.
}
