% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_date.R
\name{get_plays_by_date}
\alias{get_plays_by_date}
\title{get_plays_by_date}
\usage{
get_plays_by_date(date, shooting_plays_only = NULL, utc_offset = NULL)
}
\arguments{
\item{date}{The date to return in ISO 8601 format (YYYY-MM-DD).}

\item{shooting_plays_only}{When true, returns only shooting plays.}

\item{utc_offset}{Shifts the date range by this number of hours from UTC.}
}
\description{
Returns all recorded plays for a UTC date.
}
