% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_player_id.R
\name{get_plays_by_player_id}
\alias{get_plays_by_player_id}
\title{get_plays_by_player_id}
\usage{
get_plays_by_player_id(player_id, season, shooting_plays_only = NULL)
}
\arguments{
\item{player_id}{The player ID.}

\item{season}{The season to return.}

\item{shooting_plays_only}{When true, returns only shooting plays.}
}
\description{
Returns all recorded plays for a player and season.
}
