% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_team.R
\name{get_plays_by_team}
\alias{get_plays_by_team}
\title{get_plays_by_team}
\usage{
get_plays_by_team(season, team, shooting_plays_only = NULL)
}
\arguments{
\item{season}{The season to return.}

\item{team}{The team name to return.}

\item{shooting_plays_only}{When true, returns only shooting plays.}
}
\description{
Returns all recorded plays for a team and season.
}
