% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_tournament.R
\name{get_plays_by_tournament}
\alias{get_plays_by_tournament}
\title{get_plays_by_tournament}
\usage{
get_plays_by_tournament(tournament, season, shooting_plays_only = NULL)
}
\arguments{
\item{tournament}{The tournament to return, such as NCAA or NIT.}

\item{season}{The season to return.}

\item{shooting_plays_only}{When true, returns only shooting plays.}
}
\description{
Returns all recorded plays for a tournament and season.
}
