% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portal_transfers.R
\name{get_portal_transfers}
\alias{get_portal_transfers}
\title{get_portal_transfers}
\usage{
get_portal_transfers(
  year = NULL,
  source_team = NULL,
  destination_team = NULL,
  source_conference = NULL,
  destination_conference = NULL,
  position = NULL
)
}
\arguments{
\item{year}{Filters results to the specified transfer season.}

\item{source_team}{Filters results to the specified source team.}

\item{destination_team}{Filters results to the specified destination team.}

\item{source_conference}{Filters results to the specified source conference abbreviation.}

\item{destination_conference}{Filters results to the specified destination conference abbreviation.}

\item{position}{Filters results to the specified player position.}
}
\description{
Returns historical transfer portal activity.
}
