% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{get_rankings}
\usage{
get_rankings(
  season = NULL,
  season_type = NULL,
  week = NULL,
  poll_type = NULL,
  team = NULL,
  conference = NULL
)
}
\arguments{
\item{season}{Filters results to the specified season.}

\item{season_type}{Filters results to the specified season type.}

\item{week}{Filters results to the specified week.}

\item{poll_type}{Filters results to the AP or Coaches Poll.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns historical poll rankings.
}
