% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recruits.R
\name{get_recruits}
\alias{get_recruits}
\title{get_recruits}
\usage{
get_recruits(year = NULL, team = NULL, conference = NULL, position = NULL)
}
\arguments{
\item{year}{Filters results to the specified recruiting year.}

\item{team}{Filters results to the specified college team.}

\item{conference}{Filters results to the specified conference abbreviation.}

\item{position}{Filters results to the specified player position.}
}
\description{
Returns historical composite player recruiting rankings and ratings.
}
