% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreboard.R
\name{get_scoreboard}
\alias{get_scoreboard}
\title{get_scoreboard}
\usage{
get_scoreboard(conference = NULL)
}
\arguments{
\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns live scoreboard data. This endpoint requires Patreon Tier 1 access or higher.
}
