% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srs.R
\name{get_srs}
\alias{get_srs}
\title{get_srs}
\usage{
get_srs(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Filters results to the specified season.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns Simple Rating System (SRS) ratings.
}
