% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substitutions_by_player_id.R
\name{get_substitutions_by_player_id}
\alias{get_substitutions_by_player_id}
\title{get_substitutions_by_player_id}
\usage{
get_substitutions_by_player_id(player_id, season)
}
\arguments{
\item{player_id}{The player ID.}

\item{season}{The season to return.}
}
\description{
Returns all recorded player substitutions for a player and season.
}
