% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_leaderboard_stats.R
\name{get_team_leaderboard_stats}
\alias{get_team_leaderboard_stats}
\title{get_team_leaderboard_stats}
\usage{
get_team_leaderboard_stats(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Filters results to the specified season.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns team leaderboard statistics. This endpoint requires Patreon Tier 2 access or higher.
}
