% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_recruiting_rankings.R
\name{get_team_recruiting_rankings}
\alias{get_team_recruiting_rankings}
\title{get_team_recruiting_rankings}
\usage{
get_team_recruiting_rankings(year = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{year}{Filters results to the specified recruiting year.}

\item{team}{Filters results to the specified team name.}

\item{conference}{Filters results to the specified conference abbreviation.}
}
\description{
Returns historical composite team recruiting rankings.
}
