% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_roster.R
\name{get_team_roster}
\alias{get_team_roster}
\title{get_team_roster}
\usage{
get_team_roster(season, team = NULL)
}
\arguments{
\item{season}{The season to return.}

\item{team}{Filters results to the specified team name.}
}
\description{
Returns team rosters for a season.
}
