% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{get_teams}
\usage{
get_teams(conference = NULL, season = NULL)
}
\arguments{
\item{conference}{Filters results to the specified conference abbreviation.}

\item{season}{Returns conference membership for the specified season.}
}
\description{
Returns team and conference information.
}
