% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_cbbd}
\alias{query_cbbd}
\title{Query the College Basketball Data API}
\usage{
query_cbbd(endpoint, query = list())
}
\arguments{
\item{endpoint}{A valid CBBD API endpoint, such as \code{venues}}

\item{query}{A named list of parameters with which to query the endpoint, such as \code{list(season = 2025)}. When wishing to query an endpoint without parameters, use an empty list. Defaults to an empty list.}
}
\value{
The data from an API endpoint call, formatted as a list.
}
\description{
A low level function used to query the College Basketball Data API. For a given \code{endpoint} and an optional \code{query} list, returns the results of an API call in list format.

It is advised that \code{cbbdr} users instead use \verb{get_*} functions when available, as this function returns raw, unformatted data, and \verb{get_*} functions return cleaned data. It can be used to query new endpoints that do not have currently implemented \verb{get_*} functions available.
}
\examples{
\donttest{
try({
  query_cbbd("venues")
  query_cbbd("teams", query = list(conference = "ACC", season = 2024))
})
}
}
