% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_api_key}
\alias{set_api_key}
\title{Set CBBD API key}
\usage{
set_api_key()
}
\arguments{
\item{key}{An API key for interfacing with the College Basketball Data API, as a string. Should not include "Bearer " in front of the key.}
}
\description{
Sets the API key for a user to interface with the College Basketball Data API. Will prompt interactively if no key is passed in as a function argument. For more details about getting an API key, see https://collegebasketballdata.com/key.
}
\examples{
\donttest{
try(set_api_key())
}
}
