% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adjusted_efficiency.R
\name{get_adjusted_efficiency}
\alias{get_adjusted_efficiency}
\title{get_adjusted_efficiency}
\usage{
get_adjusted_efficiency(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Optional season filter}

\item{team}{Optional team filter}

\item{conference}{Optional conference abbreviation filter}
}
\description{
Retrieves adjusted efficiency ratings for the provided season, team, or conference.
}
