% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conference_history.R
\name{get_conference_history}
\alias{get_conference_history}
\title{get_conference_history}
\usage{
get_conference_history(conference = NULL)
}
\arguments{
\item{conference}{Optional conference abbreviation filter}
}
\description{
Retrieves historical conference membership information
}
