% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draft_picks.R
\name{get_draft_picks}
\alias{get_draft_picks}
\title{get_draft_picks}
\usage{
get_draft_picks(
  year = NULL,
  draft_team = NULL,
  source_team = NULL,
  position = NULL
)
}
\arguments{
\item{year}{Optional draft year filter}

\item{draft_team}{Optional NBA team filter}

\item{source_team}{Optional source team (e.g. NCAA) filter}

\item{position}{Optional player position abbreviation filter}
}
\description{
Retrieves historical NBA draft picks
}
