% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elo.R
\name{get_elo}
\alias{get_elo}
\title{get_elo}
\usage{
get_elo(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Optional season filter}

\item{team}{Optional team filter}

\item{conference}{Optional conference filter}
}
\description{
Retrieves historical Elo ratings
}
