% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_games}
\alias{get_games}
\title{get_games}
\usage{
get_games(
  start_date_range = NULL,
  end_date_range = NULL,
  team = NULL,
  conference = NULL,
  season = NULL,
  season_type = NULL,
  status = NULL,
  tournament = NULL
)
}
\arguments{
\item{start_date_range}{Optional start timestamp in ISO 8601 format}

\item{end_date_range}{Optional end timestamp in ISO 8601 format}

\item{team}{Optional team name filter}

\item{conference}{Optional conference abbreviation filter}

\item{season}{Optional season filter}

\item{season_type}{Optional season type filter}

\item{status}{Optional game status filter}

\item{tournament}{Optional tournament filter (e.g. NCAA, NIT, etc)}
}
\description{
Returns information on the first 3000 games that match the provided filters, ordered by start date.
}
