% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lines.R
\name{get_lines}
\alias{get_lines}
\title{get_lines}
\usage{
get_lines(
  season = NULL,
  team = NULL,
  conference = NULL,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{Optional season filter}

\item{team}{Optional team name filter}

\item{conference}{Optional conference abbreviation filter}

\item{start_date_range}{Optional start timestamp in ISO 8601 format}

\item{end_date_range}{Optional end timestamp in ISO 8601 format}
}
\description{
Returns betting lines for the first 3000 games that match the provided filters, ordered by start date.
}
