% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineup_stats_by_game.R
\name{get_lineup_stats_by_game}
\alias{get_lineup_stats_by_game}
\title{get_lineup_stats_by_game}
\usage{
get_lineup_stats_by_game(game_id)
}
\arguments{
\item{game_id}{Required game id filter}
}
\description{
Queries lineup statistics for a specific game
}
