% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineups_by_team_season.R
\name{get_lineups_by_team_season}
\alias{get_lineups_by_team_season}
\title{get_lineups_by_team_season}
\usage{
get_lineups_by_team_season(
  season,
  team,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{Required season filter}

\item{team}{Required team filter}

\item{start_date_range}{Optional start date range filter}

\item{end_date_range}{Optional end date range filter}
}
\description{
Queries lineup statistics for a given team and season
}
