% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_season_shooting_stats.R
\name{get_player_season_shooting_stats}
\alias{get_player_season_shooting_stats}
\title{get_player_season_shooting_stats}
\usage{
get_player_season_shooting_stats(
  season,
  season_type = NULL,
  team = NULL,
  conference = NULL,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{Required season filter}

\item{season_type}{Optional season type filter}

\item{team}{Team filter, required if conference is not provided}

\item{conference}{Conference abbreviation filter, required if team is not provided}

\item{start_date_range}{Optional start date range filter}

\item{end_date_range}{Optional end date range filter}
}
\description{
Retrieves player season shooting statistics
}
