% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_season_stats.R
\name{get_player_season_stats}
\alias{get_player_season_stats}
\title{get_player_season_stats}
\usage{
get_player_season_stats(
  season,
  season_type = NULL,
  team = NULL,
  conference = NULL,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{Required season filter}

\item{season_type}{Optional season type filter}

\item{team}{Optional team name filter}

\item{conference}{Optional conference abbreviation filter}

\item{start_date_range}{Required season filter}

\item{end_date_range}{Optional season type filter}
}
\description{
Returns player statistics by season
}
