% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays.R
\name{get_plays}
\alias{get_plays}
\title{get_plays}
\usage{
get_plays(game_id, shooting_plays_only = NULL)
}
\arguments{
\item{game_id}{Game id filter}

\item{shooting_plays_only}{Optional filter to only return shooting plays}
}
\description{
Returns all plays for a given game
}
