% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_player_id.R
\name{get_plays_by_player_id}
\alias{get_plays_by_player_id}
\title{get_plays_by_player_id}
\usage{
get_plays_by_player_id(player_id, season, shooting_plays_only = NULL)
}
\arguments{
\item{player_id}{Required player id filter}

\item{season}{Required season filter}

\item{shooting_plays_only}{Optional filter to only return shooting plays}
}
\description{
Retrieve all plays for a given player and season
}
