% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_team.R
\name{get_plays_by_team}
\alias{get_plays_by_team}
\title{get_plays_by_team}
\usage{
get_plays_by_team(season, team, shooting_plays_only = NULL)
}
\arguments{
\item{season}{Required season filter}

\item{team}{Required team filter}

\item{shooting_plays_only}{Optional filter to only return shooting plays}
}
\description{
Retrieve all plays for a given team and season
}
