% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays_by_tournament.R
\name{get_plays_by_tournament}
\alias{get_plays_by_tournament}
\title{get_plays_by_tournament}
\usage{
get_plays_by_tournament(tournament, season, shooting_plays_only = NULL)
}
\arguments{
\item{tournament}{Required tournament filter (e.g. NCAA, NIT, etc)}

\item{season}{Required season filter}

\item{shooting_plays_only}{Optional filter to only return shooting plays}
}
\description{
Retrieve all plays for a given tournament and season
}
