% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portal_transfers.R
\name{get_portal_transfers}
\alias{get_portal_transfers}
\title{get_portal_transfers}
\usage{
get_portal_transfers(
  year = NULL,
  source_team = NULL,
  destination_team = NULL,
  source_conference = NULL,
  destination_conference = NULL,
  position = NULL
)
}
\arguments{
\item{year}{Source team filter}

\item{source_team}{Destination team filter}

\item{destination_team}{Source conference filter}

\item{source_conference}{Destination conference filter}

\item{destination_conference}{Position filter}

\item{position}{Source team filter}
}
\description{
Retrieves historical transfer portal activity
}
