% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{get_rankings}
\usage{
get_rankings(
  season = NULL,
  season_type = NULL,
  week = NULL,
  poll_type = NULL,
  team = NULL,
  conference = NULL
)
}
\arguments{
\item{season}{Optional season filter}

\item{season_type}{Optional season type filter}

\item{week}{Optional week filter}

\item{poll_type}{Optional poll type filter ("ap" or "coaches")}

\item{team}{Optional team filter}

\item{conference}{Optional conference filter}
}
\description{
Retrieves historical poll data
}
