% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recruits.R
\name{get_recruits}
\alias{get_recruits}
\title{get_recruits}
\usage{
get_recruits(year = NULL, team = NULL, conference = NULL, position = NULL)
}
\arguments{
\item{year}{Optional year filter}

\item{team}{Optional college team filter}

\item{conference}{Optional college conference filter}

\item{position}{Optional position filter}
}
\description{
Retrieves historical composite player recruiting ranking and ratings
}
