% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoreboard.R
\name{get_scoreboard}
\alias{get_scoreboard}
\title{get_scoreboard}
\usage{
get_scoreboard(conference = NULL)
}
\arguments{
\item{conference}{Optional conference filter}
}
\description{
Retrieves live scoreboard data (requires a Patreon Tier 1 subscription or higher)
}
