% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srs.R
\name{get_srs}
\alias{get_srs}
\title{get_srs}
\usage{
get_srs(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Optional season filter}

\item{team}{Optional team filter}

\item{conference}{Optional conference abbreviation filter}
}
\description{
Retrieves SRS ratings for the provided season, team, or conference.
}
