% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substitutions_by_game.R
\name{get_substitutions_by_game}
\alias{get_substitutions_by_game}
\title{get_substitutions_by_game}
\usage{
get_substitutions_by_game(game_id)
}
\arguments{
\item{game_id}{Game id filter}
}
\description{
Returns all player substitutions for a given game
}
