% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_substitutions_by_team.R
\name{get_substitutions_by_team}
\alias{get_substitutions_by_team}
\title{get_substitutions_by_team}
\usage{
get_substitutions_by_team(season, team)
}
\arguments{
\item{season}{Required season filter}

\item{team}{Required team filter}
}
\description{
Retrieve all player substitutions for a given team and season
}
