% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_leaderboard_stats.R
\name{get_team_leaderboard_stats}
\alias{get_team_leaderboard_stats}
\title{get_team_leaderboard_stats}
\usage{
get_team_leaderboard_stats(season = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{season}{Optional season filter}

\item{team}{Optional team name filter}

\item{conference}{Optional conference abbreviation filter}
}
\description{
Returns premium team leaderboard statistics by season, team, or conference (requires a Patreon Tier 2 subscription or higher)
}
