% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_recruiting_rankings.R
\name{get_team_recruiting_rankings}
\alias{get_team_recruiting_rankings}
\title{get_team_recruiting_rankings}
\usage{
get_team_recruiting_rankings(year = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{year}{Optional year filter}

\item{team}{Optional team filter}

\item{conference}{Optional conference abbreviation filter}
}
\description{
Retrieves historical composite team recruiting rankings
}
