% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_season_stats.R
\name{get_team_season_stats}
\alias{get_team_season_stats}
\title{get_team_season_stats}
\usage{
get_team_season_stats(
  season = NULL,
  season_type = NULL,
  team = NULL,
  conference = NULL,
  start_date_range = NULL,
  end_date_range = NULL
)
}
\arguments{
\item{season}{Optional season filter, required if team is not provided}

\item{season_type}{Optional season type filter}

\item{team}{Optional team name filter, required if season is not provided}

\item{conference}{Optional conference abbreviation filter}

\item{start_date_range}{Optional season filter, required if team is not provided}

\item{end_date_range}{Optional season type filter}
}
\description{
Returns team season statistics by year or team
}
